+++ DSharpPlus.Rest/ClientCredentialsResponse.cs
@@ -1,4 +1,4 @@
-﻿namespace DSharpPlus.OAuth2
+﻿namespace DSharpPlus.Rest
 {
 	public class ClientCredentialsResponse : BaseTokenResponse
 	{
