+++ DSharpPlus.UnitTests/DSharpPlus.UnitTests.csproj
@@ -15,7 +15,6 @@
   <ItemGroup>
     <ProjectReference Include="..\DSharpPlus.CommandsNext\DSharpPlus.CommandsNext.csproj" />
     <ProjectReference Include="..\DSharpPlus.Interactivity\DSharpPlus.Interactivity.csproj" />
-    <ProjectReference Include="..\DSharpPlus.OAuth2\DSharpPlus.OAuth2.csproj" />
     <ProjectReference Include="..\DSharpPlus.Rest\DSharpPlus.Rest.csproj" />
     <ProjectReference Include="..\DSharpPlus\DSharpPlus.csproj" />
   </ItemGroup>
