+++ DSharpPlus.sln
@@ -51,8 +51,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".github", ".github", "{430C
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DSharpPlus.UnitTests", "DSharpPlus.UnitTests\DSharpPlus.UnitTests.csproj", "{7EAB8D4F-F17B-4594-8701-48CFC18E0882}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DSharpPlus.OAuth2", "DSharpPlus.OAuth2\DSharpPlus.OAuth2.csproj", "{77279A61-D35F-4944-8AC4-6CC21E8B2F85}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -97,10 +95,6 @@ Global
 		{7EAB8D4F-F17B-4594-8701-48CFC18E0882}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7EAB8D4F-F17B-4594-8701-48CFC18E0882}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7EAB8D4F-F17B-4594-8701-48CFC18E0882}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{77279A61-D35F-4944-8AC4-6CC21E8B2F85}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{77279A61-D35F-4944-8AC4-6CC21E8B2F85}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{77279A61-D35F-4944-8AC4-6CC21E8B2F85}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{77279A61-D35F-4944-8AC4-6CC21E8B2F85}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
